refactor: post와 comment 조회 시 writerId도 같이 반환, comment 조회 시 사용자 인증 정보 꺼내오는 방식 변경

+++ backend/src/main/java/com/trend_now/backend/comment/data/vo/FindCommentsResponse.java
@@ -11,7 +11,7 @@ import java.util.List;
 @AllArgsConstructor(access = AccessLevel.PUBLIC)
 public class FindCommentsResponse {
 
-    private final int totalCommentsCount;
+    private final long totalCommentsCount;
     private final int totalPageCount;
     private final List<FindAllCommentsDto> findAllCommentsDtos;
 }
